ray-cluster/requirements.txt
ray[air]
torch
torchvision
pillow
numpy
typing
scikit-image
huggingface_hub
imageio